Created room model
The plan is to enable API authentication via this model.
The room has an API Key that will be passed to all API requests.
When the API Key received matches a room then the request will be
handled and considered valid.

Eventually, all other attributes from a room will be passed down to the
actual javascript code that is execued in the haxball headless site.
This will allow us to configure little things of the room, such as the
name, the password optionally, and its privacy.

Even more, in time we'll be able to associate all matches to a given
room, allowing us to have more fine-grained stats for a given room.